backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
gitpython==3.1.43
pydantic==2.10.3
python-multipart==0.0.19
httpx==0.28.1
